cheri: Apply the tasklet.cocci script
Run the tasklet.cocci script to fix tasklet calling API.

Coccinelle can do cross-file analysis provided that all the
affected files are in a single invocation. Thus to catch all
matches of the script use the following hack:
- Determine all directories that contain potential matches,
  e.g. by grepping for file that contain the string "tasklet".
- For each directory <DIR> run
	spatch -sp-file scripts/coccinelle/cheri/tasklet.cocci <DIR>/*.[ch]

+++ drivers/atm/eni.c
@@ -1523,7 +1523,7 @@ static irqreturn_t eni_int(int irq,void *dev_id)
 }
 
 
-static void eni_tasklet(unsigned long data)
+static void eni_tasklet(uintptr_t data)
 {
 	struct atm_dev *dev = (struct atm_dev *) data;
 	struct eni_dev *eni_dev = ENI_DEV(dev);
@@ -1840,7 +1840,7 @@ static int eni_start(struct atm_dev *dev)
 	     eni_dev->vci,eni_dev->rx_dma,eni_dev->tx_dma,
 	     eni_dev->service,buf);
 	spin_lock_init(&eni_dev->lock);
-	tasklet_init(&eni_dev->task,eni_tasklet,(unsigned long) dev);
+	tasklet_init(&eni_dev->task,eni_tasklet,(uintptr_t) dev);
 	eni_dev->events = 0;
 	/* initialize memory management */
 	buffer_mem = eni_dev->mem - (buf - eni_dev->ram);
