+++ drivers/crypto/amcc/crypto4xx_core.c
@@ -1025,7 +1025,7 @@ static void crypto4xx_unregister_alg(struct crypto4xx_device *sec_dev)
 	}
 }
 
-static void crypto4xx_bh_tasklet_cb(unsigned long data)
+static void crypto4xx_bh_tasklet_cb(uintptr_t data)
 {
 	struct device *dev = (struct device *)data;
 	struct crypto4xx_core_device *core_dev = dev_get_drvdata(dev);
@@ -1372,7 +1372,7 @@ static int crypto4xx_probe(struct platform_device *ofdev)
 
 	/* Init tasklet for bottom half processing */
 	tasklet_init(&core_dev->tasklet, crypto4xx_bh_tasklet_cb,
-		     (unsigned long) dev);
+		     (uintptr_t) dev);
 
 	core_dev->dev->ce_base = devm_platform_ioremap_resource(ofdev, 0);
 	if (IS_ERR(core_dev->dev->ce_base)) {
