+++ drivers/crypto/cavium/nitrox/nitrox_isr.c
@@ -201,7 +201,7 @@ static void clear_bmi_err_intr(struct nitrox_device *ndev)
 	dev_err_ratelimited(DEV(ndev), "BMI_INT  0x%016llx\n", value);
 }
 
-static void nps_core_int_tasklet(unsigned long data)
+static void nps_core_int_tasklet(uintptr_t data)
 {
 	struct nitrox_q_vector *qvec = (void *)(uintptr_t)(data);
 	struct nitrox_device *ndev = qvec->ndev;
