+++ drivers/crypto/intel/ixp4xx/ixp4xx_crypto.c
@@ -437,7 +437,7 @@ static void irqhandler(void *_unused)
 	tasklet_schedule(&crypto_done_tasklet);
 }
 
-static void crypto_done_action(unsigned long arg)
+static void crypto_done_action(uintptr_t arg)
 {
 	int i;
