+++ drivers/hsi/clients/nokia-modem.c
@@ -35,7 +35,7 @@ struct nokia_modem_device {
 	struct hsi_client	*cmt_speech;
 };
 
-static void do_nokia_modem_rst_ind_tasklet(unsigned long data)
+static void do_nokia_modem_rst_ind_tasklet(uintptr_t data)
 {
 	struct nokia_modem_device *modem = (struct nokia_modem_device *)data;
 
@@ -154,7 +154,7 @@ static int nokia_modem_probe(struct device *dev)
 	pflags = irq_get_trigger_type(irq);
 
 	tasklet_init(&modem->nokia_modem_rst_ind_tasklet,
-			do_nokia_modem_rst_ind_tasklet, (unsigned long)modem);
+			do_nokia_modem_rst_ind_tasklet, (uintptr_t) modem);
 	err = devm_request_irq(dev, irq, nokia_modem_rst_ind_isr,
 				pflags, "modem_rst_ind", modem);
 	if (err < 0) {
