+++ drivers/hv/channel_mgmt.c
@@ -364,7 +364,7 @@ static struct vmbus_channel *alloc_channel(void)
 	INIT_LIST_HEAD(&channel->sc_list);
 
 	tasklet_init(&channel->callback_event,
-		     vmbus_on_event, (unsigned long)channel);
+		     vmbus_on_event, (uintptr_t) channel);
 
 	hv_ringbuffer_pre_init(channel);
