+++ drivers/hv/hv.c
@@ -193,7 +193,7 @@ int hv_synic_alloc(void)
 		hv_cpu = per_cpu_ptr(hv_context.cpu_context, cpu);
 
 		tasklet_init(&hv_cpu->msg_dpc,
-			     vmbus_on_msg_dpc, (unsigned long)hv_cpu);
+			     vmbus_on_msg_dpc, (uintptr_t) hv_cpu);
 
 		if (ms_hyperv.paravisor_present && hv_isolation_type_tdx()) {
 			ret = hv_alloc_page(&hv_cpu->post_msg_page,
